+++ lib/asset/computeplan.go
@@ -1,0 +1,5 @@
+package asset
+
+func (c *ComputePlan) IsTerminated() bool {
+	return c.CancelationDate != nil || c.FailureDate != nil
+}